requirements.txt
fastapi[standard]==0.112.0
uvicorn==0.30.0
streamlit==1.35.0
yfinance==0.2.40
requests==2.32.3
sqlalchemy==2.0.30
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.8.0
pydantic-settings==2.3.0
pandas==2.2.2
plotly==5.22.0